Flesh out the parser's Monad and helpers
The Parser type caries an Ast and remaining Token list, or just a
remaining Token list if no Ast was generated by a subparser. This is
used to represent the carried state.

A few helpers are defined, such as one for mismatched tokens, and one to
consume an expected token (producing no AST). Some type definitions
exist to alias the different uses of a Parser (hopefully improving
readability).