Add convenience function for list parsing
Part of the `parseE` definition requires parsing a list of arguments. A
helper function, `many` should be reusable in other areas of the parser.

There is now a `parsePlaceholder` function that will take a
`placeholder` identifier and replace it with an `AstDummy` node. This is
useful for testing, and actually allows us to generate a few basic trees
from inputs (which appears to work!).